requirements.txt
Flask
Flask-Mail
docx-mailmerge
lxml
Flask-Cors
